+++ js/utils.js
@@ -1,0 +1,14 @@
+/* eslint-disable no-unused-vars */
+const getRandomPostiveInteger = (a, b) => {
+  const min = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const max = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+
+  return Math.floor(Math.random() * (max - min + 1) + min);
+};
+
+const generatorId = () => {
+  let id = 1;
+  return () => id++;
+};
+
+export { getRandomPostiveInteger, generatorId };
